+++ Assets/Scripts/Database/Data/DailyReward/DailyReward.cs
@@ -8,12 +8,12 @@
 	public class DailyReward : ScriptableObject
 	{
 		[SerializeField] private int _beerGain = 0;
-		[SerializeField] private int _goldGain = 0;
+		[SerializeField] private ulong _goldGain = 0;
 		[SerializeField] private int _resourceGain = 0;
 		[SerializeField] private int _mithrilGain = 0;
 
 		public int BeerGain { get { return _beerGain; } }
-		public int GoldGain { get { return _goldGain; } }
+		public ulong GoldGain { get { return _goldGain; } }
 		public int ResourceGain { get { return _resourceGain; } }
 		public int MithrilGain { get { return _mithrilGain; } }
 	}
