+++ Assets/Scripts/Database/Data/FortressData/FortressData.cs
@@ -9,7 +9,7 @@
 	{
 		[SerializeField] private int _UID = 0;
 		[SerializeField] private string _name = "resourceName";
-		[SerializeField] private int _price = 1000;
+		[SerializeField] private ulong _price = 1000;
 		[SerializeField] private ResourceData _resourceToProduce = null;
 		[SerializeField] private WeaponData[] _weaponsToProduce = null;
         [SerializeField] private Sprite _fortressIcon = null;
@@ -17,7 +17,7 @@
         public Sprite FortressIcon { get { return _fortressIcon; } }
 		public int UID { get { return _UID; } }
 		public string Name { get { return _name; } }
-		public int Price { get { return _price; } }
+		public ulong Price { get { return _price; } }
 		public ResourceData ResourceToProduce { get { return _resourceToProduce; } }
 		public WeaponData[] WeaponsToProduce { get { return _weaponsToProduce; } }
 	}
