+++ Assets/Scripts/UI/ProperUIScrpits/BuildingsPanel/ForgeDisplayer.cs
@@ -102,7 +102,7 @@
             if (_isGoldTrans)
             {
                 bool isMax;
-                int price = _converter.ComputeUpgradeCost(DatabaseManager.Instance.ForgeUpgrades.WorkerAmount, currentFortress.UForgeWorkerNbIndex);
+                ulong price = _converter.ComputeUpgradeCost(DatabaseManager.Instance.ForgeUpgrades.WorkerAmount, currentFortress.UForgeWorkerNbIndex);
 			    _workerUpgrade.Init(uData.WorkerAmount.name, uData.WorkerAmount.desc, currentFortress.UForgeWorkerNbIndex, price, _playerProfile.Gold);
 
 			    price = _converter.ComputeUpgradeCost(DatabaseManager.Instance.ForgeUpgrades.WByWorker, currentFortress.UForgeWByWorkerIndex);
