+++ Assets/Scripts/UI/ProperUIScrpits/BuildingsPanel/TradingPostDisplayer.cs
@@ -100,7 +100,7 @@
             if (_isGoldTrans)
             {
                 bool isMax = currentFortress.UForgeCycleDurationIndex >= uData.CycleDuration.max;
-                int price = _converter.ComputeUpgradeCost(DatabaseManager.Instance.TradingPostUpgrades.WorkerAmount, currentFortress.UTPWorkerNbIndex);
+                ulong price = _converter.ComputeUpgradeCost(DatabaseManager.Instance.TradingPostUpgrades.WorkerAmount, currentFortress.UTPWorkerNbIndex);
                 _workerUpgrade.Init(uData.WorkerAmount.name, uData.WorkerAmount.desc, currentFortress.UTPWorkerNbIndex, price, _playerProfile.Gold);
 
                 price = _converter.ComputeUpgradeCost(DatabaseManager.Instance.TradingPostUpgrades.SellByWorker, currentFortress.UTPSellByWorkerIndex);
